pyproject.toml
[project]
name = "model-trys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.8.1",
    "datasets>=3.6.0",
    "evaluate>=0.4.4",
    "marimo>=0.14.10",
    "marker-pdf>=1.8.1",
    "pre-commit>=4.2.0",
    "ruff>=0.12.2",
    "timm>=1.0.16",
    "torch>=2.7.1",
    "transformers>=4.53.1",
]

[tool.pytest.ini_options]
pythonpath = ["./src"]
testpaths = ["tests", "marimo_template"]

[tool.marimo.ai]
rules = "- prefer polars over pandas\n- make charts using altair"

[tool.marimo.save]
autosave = "after_delay"
autosave_delay = 1000
format_on_save = true

[tool.marimo.package_management]
manager = "uv"

[tool.marimo.runtime]
pythonpath = ["./src"]
